requirements.txt
crewai==0.28.8
crewai-tools==0.2.6
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
openai>=1.13.3
pyyaml==6.0.1
requests==2.31.0
resend==0.8.0